separated descriptoin from watchplayer
